feat: package refactoring
