+++ src/main/java/nextstep/api/auth/domain/UserDetails.java
@@ -1,4 +1,4 @@
-package nextstep.api.user;
+package nextstep.api.auth.domain;
 
 /**
  * @author : Rene Choi
